% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{app}
\alias{app}
\title{Run shiny app}
\usage{
app(display.mode = "normal", ...)
}
\arguments{
\item{display.mode}{display mode}

\item{\dots}{see \code{\link[shiny]{runApp}}}
}
\description{
Run shiny app
}
\examples{
\dontrun{
  app()
}
}
