% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweetstorm.R
\name{embed_tweet}
\alias{embed_tweet}
\title{Embed tweet}
\usage{
embed_tweet(id)
}
\arguments{
\item{id}{status id of the tweet to embed}
}
\value{
html embeddable tweet, e.g. inside a [shiny::renderUI()]
}
\description{
Embed tweet
}
\examples{
embed_tweet( "266031293945503744" )

}
