% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweetstorm.R
\name{extract_emojis}
\alias{extract_emojis}
\title{Extract emojis}
\usage{
extract_emojis(text)
}
\arguments{
\item{text}{text with emojis}
}
\value{
a tibble with the columns `Emoji` and `n`
}
\description{
Extract emojis
}
\references{
Inspired from [](http://seankross.com/2017/05/30/Which-Emojis-Does-Lucy-Use-in-Commit-Messages.html)
}
