% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweetstorm.R
\name{extract_emojis_users}
\alias{extract_emojis_users}
\title{Extract information about users that use emojis}
\usage{
extract_emojis_users(tweets)
}
\arguments{
\item{tweets}{tweets data set}
}
\description{
Extract information about users that use emojis
}
