% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweetstorm.R
\name{most_popular}
\alias{most_popular}
\alias{most_retweeted}
\alias{most_recent}
\title{most popular tweets}
\usage{
most_popular(tweets, n = 6)

most_retweeted(tweets, n = 6)

most_recent(tweets, n = 6)
}
\arguments{
\item{tweets}{tweets data}

\item{n}{number of tweets to extract}
}
\description{
most popular tweets
}
