% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweetstorm.R
\name{random_tweet}
\alias{random_tweet}
\title{random tweet}
\usage{
random_tweet(query = "#rstats")
}
\arguments{
\item{query}{twitter query, see [rtweet::search_tweets]}
}
\description{
random tweet
}
\examples{
\dontrun{
  embed_tweet( random_tweet("#rstats") )
}
}
