% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweetstorm.R
\name{summarise_hashtags}
\alias{summarise_hashtags}
\title{Summarise hashtags}
\usage{
summarise_hashtags(hashtags)
}
\arguments{
\item{hashtags}{vector of hashtags}
}
\value{
a tibble
}
\description{
Summarise hashtags
}
