% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{useR2017}
\alias{useR2017}
\title{Tweets from the useR 2017 conference}
\format{An object of class \code{data.frame} with 2003 rows and 35 columns.}
\usage{
useR2017
}
\description{
Tweets from the useR 2017 conference
}
\keyword{datasets}
