% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweetstorm.R
\name{users_datatable}
\alias{users_datatable}
\title{organise users in a data table}
\usage{
users_datatable(data)
}
\arguments{
\item{data}{tibble with user information}
}
\description{
organise users in a data table
}
